refine extensions
